refactor: refactor get_color
Introduce a `ColorFormat` Enum, and refactor `get_color` to take that as a new `color_mode` param. Default this new param to be `ColorFormat.RGB` to match the existing default behaviour of `color_format`.

Change If `color_format` is passed, raise a `DeprecationWarning`, and attempt to convert the passed value to a `ColorFormat` enum instance and use that as the value for `color_mode`