chore: Apply Black formatting
Apply Black formatting to the entire codebase.

Note that this touches _everything_, so if you're interested in running `git blame` to see who made chages to files prior to this, it will muddy the water.

In the following commit, a `.git-blame-ignore-revs` file will be added, which can be used to configure git to ignore this commit when  running git blame. see https://git-scm.com/docs/git-config#Documentation/git-config.txt-blameignoreRevsFile for details of how to enable this behaviour.